requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.4.0

# CORS middleware
starlette>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0  # For async SQLite
psycopg2-binary>=2.9.9  # For PostgreSQL support when deploying
asyncpg>=0.28.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dateutil>=2.8.2

# LLM integration
langchain-google-genai>=0.0.5
langchain-core>=0.1.0
google-generativeai>=0.3.0

# Admin interface
flask>=2.3.0
flask-admin>=1.6.1

# Environment variables
python-dotenv>=1.0.0